# web/openapi.yaml
openapi: 3.0.3
info:
  title: Helix REST API (runs subset)
  version: "1.0.0"
  description: >
    Hand-maintained subset of the orchestrator REST API used by the M4 dashboard.
    Source of truth for web/lib/api-types.ts (generated via `npm run gen:types`).
    Mirrors the Go JSON shapes in internal/store and internal/api. Expand as the
    dashboard grows; replace with an orchestrator-emitted schema when available.
servers:
  - url: /api/v1
paths:
  /runs:
    get:
      operationId: listRuns
      summary: List runs, optionally filtered by status.
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/RunStatus"
      responses:
        "200":
          description: Runs (most recent first, up to 100).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
  /runs/{run_id}:
    get:
      operationId: getRun
      summary: Fetch a run including its task tree.
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run with tasks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunDetail"
        "404":
          description: Run not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      summary: Cancel a run. Idempotent.
      parameters:
        - in: path
          name: run_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancellation acknowledged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelResponse"
components:
  schemas:
    RunStatus:
      type: string
      enum: [pending, running, succeeded, failed, cancelled]
    TaskStatus:
      type: string
      enum: [pending, ready, running, succeeded, failed, cancelled, dead]
    Run:
      type: object
      required: [id, workflow_id, status, submitted_at, submitted_by, trace_id]
      properties:
        id:
          type: string
        workflow_id:
          type: string
        status:
          $ref: "#/components/schemas/RunStatus"
        input:
          type: string
          description: Raw JSON bytes (base64 in transit).
        output:
          type: string
        error:
          type: string
        submitted_at:
          type: string
          format: date-time
        submitted_by:
          type: string
        trace_id:
          type: string
        task_ids:
          type: array
          items:
            type: string
    Task:
      type: object
      required: [id, run_id, node_id, status, attempts]
      properties:
        id:
          type: string
        run_id:
          type: string
        node_id:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        attempts:
          type: integer
    RunDetail:
      allOf:
        - $ref: "#/components/schemas/Run"
        - type: object
          required: [tasks]
          properties:
            tasks:
              type: array
              items:
                $ref: "#/components/schemas/Task"
    CancelResponse:
      type: object
      properties:
        status:
          type: string
        run_id:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string